requirements.txt
python-dotenv==1.1.1
Requests==2.32.4
streamlit==1.46.1
